adding support for duplicate pairs

+++ find_keltner_channel.py
@@ -39,13 +39,21 @@ def calculate_keltner_channels(df):
     return df
 
 
-def find_mirror_levels_in_database():
+def find_mirror_levels_in_database_and_add_kc_to_db(async_var):
     start_time = time.time ()
     connection_to_usdt_trading_pairs_ohlcv = \
         sqlite3.connect ( os.path.join ( os.getcwd () ,
                                          "datasets" ,
                                          "sql_databases" ,
                                          "all_exchanges_multiple_tables_historical_data_for_usdt_trading_pairs.db" ) )
+
+    if async_var==True:
+        connection_to_usdt_trading_pairs_ohlcv = \
+            sqlite3.connect ( os.path.join ( os.getcwd () ,
+                                             "datasets" ,
+                                             "sql_databases" ,
+                                             "async_all_exchanges_multiple_tables_historical_data_for_usdt_trading_pairs.db" ) )
+        print("got into if async_var==True")
     cursor=connection_to_usdt_trading_pairs_ohlcv.cursor()
     cursor.execute ( "SELECT name FROM sqlite_master WHERE type='table';" )
 
@@ -54,7 +62,11 @@ def find_mirror_levels_in_database():
                                          "sql_databases" ,
                                          "multiple_tables_historical_data_for_usdt_trading_pairs_with_mirror_levels_and_kc.db" )
 
-    os.remove(path_to_ohlcv_db_with_kc)
+    if os.path.exists(path_to_ohlcv_db_with_kc):
+        os.remove(path_to_ohlcv_db_with_kc)
+    print ( "before removal of db" )
+    #time.sleep(20)
+    print ( "after removal of db" )
     create_empty_database(path_to_ohlcv_db_with_kc)
 
 
@@ -74,6 +86,17 @@ def find_mirror_levels_in_database():
 
     connection_to_btc_and_usdt_trading_pairs = \
         sqlite3.connect ( path_to_db_with_USDT_and_btc_pairs )
+
+    # path_to_test_db = os.path.join ( os.getcwd () , "datasets" ,
+    #                                                     "sql_databases" ,
+    #                                                     "test_db.db" )
+
+    # if os.path.exists(path_to_test_db):
+    #     os.remove(path_to_test_db)
+
+    # connection_to_test_db = \
+    #     sqlite3.connect ( path_to_test_db )
+
     mirror_level_df = pd.DataFrame ( columns = ['USDT_pair' , 'exchange' , 'mirror_level' ,
                                                 'timestamp_for_low' , 'timestamp_for_high' ,
                                                 'low' , 'high' , 'open_time_of_candle_with_legit_low' ,
@@ -85,7 +108,7 @@ def find_mirror_levels_in_database():
         mirror_df = \
             pd.read_sql_query ( f'''select * from mirror_levels_without_duplicates''' ,
                                 connection_to_btc_and_usdt_trading_pairs )
-        print("mirror_df\n",mirror_df.to_string())
+        print("mirror_df\n",mirror_df)
         for row in range(0,len(mirror_df)):
             joint_string_for_table_name=mirror_df.loc[row,"USDT_pair"]+'_on_'+mirror_df.loc[row,"exchange"]
             print("row=",row)
@@ -119,28 +142,48 @@ def find_mirror_levels_in_database():
             data_df=\
                 pd.read_sql_query(f'''select * from "{table_in_db}"''' ,
                                   connection_to_usdt_trading_pairs_ohlcv)
-            print ( "data_df\n",data_df )
+            # data_df.to_sql(f'''{table_in_db}''',connection_to_test_db)
+            print ( "++"*80 )
+            print ( "++" * 80 )
+            print ( "++" * 80 )
+            print ( f"data_df_for_{table_in_db}\n",data_df )
 
             #data_df.set_index("Timestamp")
             #print ( "data_df\n" , data_df )
             print("---------------------------")
-            print(f'{table_in_db} is number {counter} out of {len(list_of_table_names)}\n' )
+            print(f'{table_in_db} is this number {counter} out of {len(list_of_table_names)}\n' )
 
             usdt_pair=data_df.loc[0,"trading_pair"]
             exchange=data_df.loc[0,"exchange"]
 
             #get data frame with ohlcv
             data_df.set_index("Timestamp", inplace = True)
+            pair_and_exchange_from_data_df=usdt_pair+exchange
             ##############################
+
+
 
             ################################
             data_df_with_kc=calculate_keltner_channels ( data_df )
+            # print(data_df_with_kc.columns)
+            # print ( "data_df_with_kc.loc[:,'trading_pair'].iloc[0]" ,
+            #         data_df_with_kc.loc[:,"trading_pair"].iloc[0] )
+            print(f"data_df_with_kc_for_{usdt_pair}_on_{exchange}\n",data_df_with_kc)
+            pair_and_exchange_from_data_df_with_kc=\
+                data_df_with_kc.loc[:,"trading_pair"].iloc[0]+\
+                data_df_with_kc.loc[:,"exchange"].iloc[0]
+            print(pair_and_exchange_from_data_df,"\n",
+                  pair_and_exchange_from_data_df_with_kc,"\n")
+
+            print(f"counter={counter} ")
+            if not (pair_and_exchange_from_data_df == pair_and_exchange_from_data_df_with_kc):
+                print(f"pair is {pair_and_exchange_from_data_df_with_kc}")
             data_df_with_kc.to_sql(f'''{table_in_db}''',
                                    connection_to_usdt_trading_pairs_ohlcv_with_kc,
-                                   if_exists = "replace")
+                                   if_exists = "append")
 
 
-            print("data_df_with_kc\n",data_df_with_kc.to_string() )
+            #print("data_df_with_kc\n",data_df_with_kc )
         except Exception as e:
             print(f"problem with {table_in_db}", e)
             traceback.print_exc ()
@@ -148,6 +191,7 @@ def find_mirror_levels_in_database():
     connection_to_usdt_trading_pairs_ohlcv.close()
     connection_to_btc_and_usdt_trading_pairs.close()
     connection_to_usdt_trading_pairs_ohlcv_with_kc.close()
+    # connection_to_test_db.close()
     end_time = time.time ()
     overall_time = end_time - start_time
     print ( 'overall time in minutes=' , overall_time / 60.0 )
@@ -162,5 +206,6 @@ def find_mirror_levels_in_database():
     local_time_end = dt.datetime.fromtimestamp ( unix_timestamp_end , local_timezone )
     print ( 'local_time_start=' , local_time_start )
     print ( 'local_time_end=' , local_time_end )
-
-find_mirror_levels_in_database()
+if __name__=="__main__":
+    async_var=True
+    find_mirror_levels_in_database_and_add_kc_to_db(async_var)
